+++ ui/src/index.ts
@@ -39,7 +39,13 @@ export {
   type FloatingToolbarPosition,
   type FloatingToolbarProps,
 } from './floating-toolbar.js';
-export { type DividerSides, List, type ListProps, type Sides } from './list.js';
+export {
+  type DividerSides,
+  List,
+  type ListProps,
+  type ListVerticalAlignment,
+  type Sides,
+} from './list.js';
 export { ListActionRow, type ListActionRowProps } from './list-action-row.js';
 export { ListHeader, type ListHeaderProps } from './list-header.js';
 export {